reset structure for rect: add type property, no longer assign to block
